+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Wang 2025 rivaroxaban ([doi:10.3389/fphar.2025.1574949](https://doi.org/10.3389/fphar.2025.1574949)) - Chinese adults with non-valvular atrial fibrillation.
+- Add Garcia 2025 garadacimab ([doi:10.1002/psp4.70009](https://doi.org/10.1002/psp4.70009)) -- healthy volunteers and adolescents and adults with hereditary angioedema.
 
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.